moved MainActivity and NoteFragment files into ui package
